avoid using private forks of ctapipe and ctapipe-extra
by replacing the forks, by their specific commits we clearly state
that we need these specific commits for digicampipe to work.

However we also clearly state, that we did not add anything to these
projects, we just froze them in time.

This is basically handled by version numbers, but ctapipe seems to
publish not too often.